requirements.txt
alabaster==0.7.12
appdirs==1.4.4
appnope==0.1.0
argon2-cffi==19.2.0
asgiref==3.2.7
astroid==2.4.1
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
black==19.10b0
boto3==1.14.62
botocore==1.17.62
certifi==2020.4.5.1
cffi==1.14.0
cfgv==3.1.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
configparser==5.0.0
coverage==5.1
crayons==0.3.1
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.7
django-allauth==0.41.0
django-anymail==7.1.0
django-coverage-plugin==1.8.0
django-crispy-forms==1.9.0
django-debug-toolbar==2.2
django-environ==0.4.5
django-extensions==2.2.9
django-heroku==0.3.1
django-model-utils==4.0.0
django-redis==4.11.0
django-selenium-login==2.0.0
django-storages==1.10.1
django-stubs==1.5.0
docutils==0.15.2
environ==1.0
factory-boy==2.12.0
Faker==4.1.0
filelock==3.0.12
flake8==3.8.2
flake8-isort==3.0.0
gunicorn==20.0.4
hiredis==1.0.1
identify==1.4.19
idna==2.9
imagesize==1.2.0
ipdb==0.13.2
ipython==7.15.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.3.0
mypy==0.770
mypy-extensions==0.4.3
nodeenv==1.4.0
oauthlib==3.1.0
packaging==20.4
parso==0.7.0
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pluggy==0.13.1
pre-commit==2.4.0
prompt-toolkit==3.0.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
py==1.8.1
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.6.1
pylint==2.5.2
pylint-django==2.0.15
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==5.4.2
pytest-django==3.9.0
pytest-sugar==0.9.3
python-dateutil==2.8.1
python-slugify==4.0.0
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.0
regex==2020.5.14
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
selenium==3.141.0
sentry-sdk==0.14.4
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.0.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
termcolor==1.1.0
testfixtures==6.14.1
text-unidecode==1.3
toml==0.10.1
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9
virtualenv==20.0.21
wcwidth==0.2.3
webdriver-manager==3.2.1
Werkzeug==1.0.1
whitenoise==5.0.1
wrapt==1.12.1
